Split worker into separate module
