+++ datafusion-scheduler/src/lib.rs
@@ -15,22 +15,23 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::query::{Query, WorkItem};
-use crossbeam_deque::Steal;
+use std::sync::Arc;
+
+use futures::stream::{BoxStream, StreamExt};
+
 use datafusion::arrow::error::Result as ArrowResult;
 use datafusion::arrow::record_batch::RecordBatch;
 use datafusion::error::Result;
 use datafusion::execution::context::TaskContext;
 use datafusion::physical_plan::ExecutionPlan;
-use futures::stream::{BoxStream, StreamExt};
-use parking_lot::{Condvar, Mutex};
-use std::cell::RefCell;
-use std::sync::Arc;
-use std::thread::JoinHandle;
+
+use crate::query::{Query, WorkItem};
+use crate::worker::WorkerPool;
 
 mod node;
 mod query;
 mod repartition;
+mod worker;
 
 pub struct Scheduler {
     pool: WorkerPool,
@@ -60,213 +61,22 @@ impl Drop for Scheduler {
     }
 }
 
-#[derive(Debug)]
-struct WorkerThread {
-    handle: JoinHandle<()>,
-}
-
-struct ThreadState {
-    idx: usize,
-    shared: Arc<SharedState>,
-    local: crossbeam_deque::Worker<WorkItem>,
-    steal: Vec<(usize, crossbeam_deque::Stealer<WorkItem>)>,
-}
-
-thread_local! {
-    static CURRENT_THREAD: RefCell<Option<ThreadState>> = RefCell::new(None)
-}
-
-/// Spawns a [`WorkItem`] to the worker local queue
-fn spawn_local(item: WorkItem) {
-    ThreadState::local(|s| {
-        println!("Spawning {:?} to local worker {}", item, s.idx);
-        s.local.push(item)
-    })
-}
-
-impl ThreadState {
-    fn find_task(&self) -> Option<WorkItem> {
-        if let Some(item) = self.local.pop() {
-            println!("Worker {} fetched {:?} from local queue", self.idx, item);
-            return Some(item);
-        }
-
-        loop {
-            match self.shared.injector.steal_batch_and_pop(&self.local) {
-                Steal::Success(item) => {
-                    println!("Worker {} fetched {:?} from injector", self.idx, item);
-                    return Some(item);
-                }
-                Steal::Retry => continue,
-                Steal::Empty => {}
-            }
-
-            // TODO: Should the search order be randomized?
-            for (idx, steal) in &self.steal {
-                match steal.steal() {
-                    Steal::Success(item) => {
-                        println!(
-                            "Worker {} stole {:?} from worker {}",
-                            self.idx, item, idx
-                        );
-                        return Some(item);
-                    }
-                    Steal::Retry => continue,
-                    Steal::Empty => {}
-                }
-            }
-
-            println!("Worker {} failed to find any work", self.idx);
-            return None;
-        }
-    }
-
-    fn local<F, T>(f: F) -> T
-    where
-        F: FnOnce(&Self) -> T,
-    {
-        CURRENT_THREAD.with(|s| f(s.borrow().as_ref().expect("worker thread")))
-    }
-
-    fn run(self) {
-        let shared = self.shared.clone();
-        let worker_idx = self.idx;
-
-        CURRENT_THREAD.with(|s| *s.borrow_mut() = Some(self));
-
-        loop {
-            match Self::local(Self::find_task) {
-                Some(task) => task.do_work(),
-                None => {
-                    if !Self::wait_for_input(worker_idx, &shared) {
-                        println!("Worker {} shutting down", worker_idx);
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    /// Park this thread, returning false if this worker should terminate
-    fn wait_for_input(worker_idx: usize, shared: &SharedState) -> bool {
-        let mut lock = shared.inner.lock();
-        if lock.is_shutdown {
-            return false;
-        }
-
-        println!("Worker {} entered sleep", worker_idx);
-        shared.signal.wait(&mut lock);
-        println!("Worker {} exited sleep", worker_idx);
-
-        !lock.is_shutdown
-    }
-}
-
-#[derive(Debug)]
-struct SharedState {
-    injector: crossbeam_deque::Injector<WorkItem>,
-    signal: Condvar,
-    inner: Arc<Mutex<SharedStateInner>>,
-}
-
-#[derive(Debug)]
-struct SharedStateInner {
-    is_shutdown: bool,
-}
-
-#[derive(Debug)]
-struct WorkerPool {
-    shared: Arc<SharedState>,
-    workers: Vec<WorkerThread>,
-}
-
-impl WorkerPool {
-    fn new(workers: usize) -> Self {
-        let shared = Arc::new(SharedState {
-            injector: crossbeam_deque::Injector::new(),
-            signal: Default::default(),
-            inner: Arc::new(Mutex::new(SharedStateInner { is_shutdown: false })),
-        });
-
-        let worker_queues: Vec<_> = (0..workers)
-            .map(|_| crossbeam_deque::Worker::new_fifo())
-            .collect();
-        let stealers: Vec<_> = worker_queues.iter().map(|x| x.stealer()).collect();
-
-        let workers = worker_queues
-            .into_iter()
-            .enumerate()
-            .map(|(idx, local)| {
-                let steal = stealers
-                    .iter()
-                    .enumerate()
-                    .filter(|(steal_idx, _)| *steal_idx != idx)
-                    .map(|(steal_idx, steal)| (steal_idx, steal.clone()))
-                    .collect();
-
-                let state = ThreadState {
-                    idx,
-                    local,
-                    steal,
-                    shared: shared.clone(),
-                };
-
-                WorkerThread {
-                    handle: std::thread::spawn(move || state.run()),
-                }
-            })
-            .collect();
-
-        WorkerPool { shared, workers }
-    }
-
-    fn spawner(&self) -> Spawner {
-        Spawner {
-            shared: self.shared.clone(),
-        }
-    }
-
-    fn shutdown(&mut self) {
-        println!("Shutting down worker pool");
-        self.shared.inner.lock().is_shutdown = true;
-        self.shared.signal.notify_all();
-
-        for (idx, worker) in self.workers.drain(..).enumerate() {
-            if let Err(_) = worker.handle.join() {
-                panic!("worker {} panicked", idx)
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Spawner {
-    shared: Arc<SharedState>,
-}
-
-impl Spawner {
-    fn spawn(&self, task: WorkItem) {
-        println!("Spawning {:?} to any worker", task);
-        self.shared.injector.push(task);
-
-        // Ensure that at least one worker thread is awake
-        self.shared.signal.notify_one();
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use std::ops::Range;
+
+    use futures::TryStreamExt;
+    use rand::{Rng, thread_rng};
+    use rand::distributions::uniform::SampleUniform;
+
     use datafusion::arrow::array::{ArrayRef, PrimitiveArray};
     use datafusion::arrow::datatypes::{ArrowPrimitiveType, Float64Type, Int32Type};
     use datafusion::arrow::record_batch::RecordBatch;
     use datafusion::datasource::MemTable;
     use datafusion::physical_plan::displayable;
     use datafusion::prelude::{SessionConfig, SessionContext};
-    use futures::TryStreamExt;
-    use rand::distributions::uniform::SampleUniform;
-    use rand::{thread_rng, Rng};
-    use std::ops::Range;
+
+    use super::*;
 
     fn generate_primitive<T, R>(
         rng: &mut R,
